wip: release lib: Factor out Ash.SatSolver into independent library
