+++ lib/ash/policy/authorizer/verifiers/verify_sat_solver_implementation.ex
@@ -7,7 +7,7 @@ defmodule Ash.Policy.Authorizer.Verifiers.VerifySatSolverImplementation do
   use Spark.Dsl.Verifier
 
   def verify(_dsl) do
-    Ash.SatSolver.Implementation.check!()
+    Crux.Implementation.check!()
     :ok
   end
 end
