+++ lib/ash/policy/check/action.ex
@@ -6,7 +6,7 @@ defmodule Ash.Policy.Check.Action do
   @moduledoc "This check is true when the action name matches the provided action name."
   use Ash.Policy.SimpleCheck
 
-  import Ash.SatSolver.Expression, only: [b: 1]
+  import Crux.Expression, only: [b: 1]
 
   @impl true
   def describe(options) do
