+++ lib/ash/policy/check/action_type.ex
@@ -6,7 +6,7 @@ defmodule Ash.Policy.Check.ActionType do
   @moduledoc "This check is true when the action type matches the provided type"
   use Ash.Policy.SimpleCheck
 
-  import Ash.SatSolver.Expression, only: [b: 1]
+  import Crux.Expression, only: [b: 1]
 
   @impl true
   def describe(options) do
