+++ test/policy/policy_test.exs
@@ -5,7 +5,7 @@
 defmodule Ash.Test.Policy.Policy do
   use ExUnit.Case, async: true
 
-  import Ash.SatSolver.Expression, only: [b: 1]
+  import Crux.Expression, only: [b: 1]
 
   defmodule RuntimeCheck do
     @moduledoc false
